+++ src/types.rs
@@ -13,10 +13,10 @@ pub struct VertexCoord{pub v: Vector3<f32>}
 pub struct ScreenPos{pub v: Vector2<i32>}
 
 #[derive(Copy, Clone, Debug)]
-pub struct Time{pub n: f32}
+pub struct Time{pub n: f64}
 
 impl Time {
     pub fn now() -> Time {
-        Time{n: time::precise_time_ns() as f32 / 1_000_000_000.0}
+        Time{n: time::precise_time_ns() as f64 / 1_000_000_000.0}
     }
 }
